+++ src/components/formularios/index.ts
@@ -1,2 +1,3 @@
 export { default as AfectadosForm } from "./AfectadosForm.vue";
 export { default as VinculadosForm } from "./VinculadosForm.vue";
+export { default as EfectosForm } from "./EfectosForm.vue";
